Fusion de la branche hitbox avec develop Modification des IntegerProperty en FloatProperty pour pouvoir mieux influencer la vitesse de déplacement

+++ src/main/java/application/controleur/listeners/EnvironnementListener.java
@@ -16,7 +16,7 @@ public class EnvironnementListener {
             public void onChanged(Change<? extends Materiau> change) {
                 while (change.next()) {
                     if (change.wasRemoved()) {
-                        envVue.supprimerBloc(change.getRemoved().get(0).getY() * WIDTH + change.getRemoved().get(0).getX());
+                        envVue.supprimerBloc((int)change.getRemoved().get(0).getY() * WIDTH + (int)change.getRemoved().get(0).getX());
                     }
                 }
             }
@@ -26,7 +26,7 @@ public class EnvironnementListener {
             public void onChanged(Change<? extends Arbre> change) {
                 while (change.next()) {
                     if (change.wasRemoved()) {
-                        envVue.supprimerArbre(change.getRemoved().get(0).getY() * WIDTH + change.getRemoved().get(0).getX());
+                        envVue.supprimerArbre((int)change.getRemoved().get(0).getY() * WIDTH + (int)change.getRemoved().get(0).getX());
                     }
                 }
             }
